Split out TOK and Token into tokens.h and tokens.c
