+++ src/posix.mak
@@ -136,7 +136,8 @@ DMD_OBJS = \
 	builtin.o ctfeexpr.o clone.o aliasthis.o \
 	arrayop.o json.o unittests.o \
 	imphint.o argtypes.o apply.o sapply.o sideeffect.o \
-	intrange.o canthrow.o target.o nspace.o errors.o
+	intrange.o canthrow.o target.o nspace.o errors.o \
+	tokens.o
 
 ROOT_OBJS = \
 	rmem.o port.o man.o stringtable.o response.o \
@@ -194,7 +195,8 @@ SRC = win32.mak posix.mak osmodel.mak \
 	argtypes.c apply.c sapply.c sideeffect.c \
 	intrange.h intrange.c canthrow.c target.c target.h \
 	scanmscoff.c scanomf.c ctfe.h ctfeexpr.c \
-	ctfe.h ctfeexpr.c visitor.h nspace.h nspace.c
+	ctfe.h ctfeexpr.c visitor.h nspace.h nspace.c \
+	tokens.h tokens.c
 
 ROOT_SRC = $(ROOT)/root.h \
 	$(ROOT)/array.h \
